refactor: rename CoWShedWidget
